Code changes to rename files.
